requirements.txt
google-adk
google-cloud-bigquery
fastapi
uvicorn
websockets
python-dotenv 